@@ -350,7 +350,7 @@ def plot_features_in_2d(
     values: Float[Tensor, "timesteps d_hidden feats"],
     colors: Optional[list[Any]] = None,  # shape [timesteps feats]
     title: Optional[str | list[str]] = None,
-    subplot_titles: Optional[list[str] | list[list[str]]] = None,
+    subplot_titles: Optional[list[str] | list[list[str]]] = None,  # type: ignore
     save: Optional[str] = None,
     colab: bool = False,
 ):
@@ -435,7 +435,7 @@ def plot_features_in_2d(
         if title:
             fig.suptitle(title[t], fontsize=15)
         if subplot_titles:
-            ax.set_title(subplot_titles[t], fontsize=12)
+            ax.set_title(subplot_titles[t], fontsize=12)  # type: ignore
         fig.canvas.draw_idle()
 
     def play(event: Any):
