@@ -1,4 +1,4 @@
-__version__ = "3.2.3"
+__version__ = "3.21.1"
 
 
 from .analysis.hooked_sae_transformer import HookedSAETransformer
@@ -9,11 +9,14 @@ from .config import (
     PretokenizeRunnerConfig,
 )
 from .evals import run_evals
-from .pretokenize_runner import pretokenize_runner
+from .pretokenize_runner import PretokenizeRunner, pretokenize_runner
 from .sae import SAE, SAEConfig, TopKActivation
+
+
 from .sae_training_runner import SAETrainingRunner
 from .training.activations_store import ActivationsStore
 from .training.training_sae import TrainingSAE, TrainingSAEConfig
+from .training.upload_saes_to_huggingface import upload_saes_to_huggingface
 
 __all__ = [
     "SAE",
@@ -27,6 +30,8 @@ __all__ = [
     "CacheActivationsRunnerConfig",
     "CacheActivationsRunner",
     "PretokenizeRunnerConfig",
+    "PretokenizeRunner",
     "pretokenize_runner",
     "run_evals",
+    "upload_saes_to_huggingface",
 ]
