@@ -75,13 +75,14 @@ class HookedSAETransformer(HookedTransformer):
         super().__init__(*model_args, **model_kwargs)
         self.acts_to_saes: Dict[str, SAE] = {}
 
-    def add_sae(self, sae: SAE):
+    def add_sae(self, sae: SAE, use_error_term: Optional[bool] = None):
         """Attaches an SAE to the model
 
         WARNING: This sae will be permanantly attached until you remove it with reset_saes. This function will also overwrite any existing SAE attached to the same hook point.
 
         Args:
             sae: SparseAutoencoderBase. The SAE to attach to the model
+            use_error_term: (Optional[bool]) If provided, will set the use_error_term attribute of the SAE to this value. Determines whether the SAE returns input or reconstruction. Defaults to None.
         """
         act_name = sae.cfg.hook_name
         if (act_name not in self.acts_to_saes) and (act_name not in self.hook_dict):
@@ -90,6 +91,10 @@ class HookedSAETransformer(HookedTransformer):
             )
             return
 
+        if use_error_term is not None:
+            if not hasattr(sae, "_original_use_error_term"):
+                sae._original_use_error_term = sae.use_error_term  # type: ignore
+            sae.use_error_term = use_error_term
         self.acts_to_saes[act_name] = sae
         set_deep_attr(self, act_name, sae)
         self.setup()
@@ -110,6 +115,11 @@ class HookedSAETransformer(HookedTransformer):
                 f"No SAE is attached to {act_name}. There's nothing to reset."
             )
             return
+
+        current_sae = self.acts_to_saes[act_name]
+        if hasattr(current_sae, "_original_use_error_term"):
+            current_sae.use_error_term = current_sae._original_use_error_term
+            delattr(current_sae, "_original_use_error_term")
 
         if prev_sae:
             set_deep_attr(self, act_name, prev_sae)
@@ -154,6 +164,7 @@ class HookedSAETransformer(HookedTransformer):
         *model_args: Any,
         saes: Union[SAE, List[SAE]] = [],
         reset_saes_end: bool = True,
+        use_error_term: Optional[bool] = None,
         **model_kwargs: Any,
     ) -> Union[
         None,
@@ -169,9 +180,12 @@ class HookedSAETransformer(HookedTransformer):
             *model_args: Positional arguments for the model forward pass
             saes: (Union[HookedSAE, List[HookedSAE]]) The SAEs to be attached for this forward pass
             reset_saes_end (bool): If True, all SAEs added during this run are removed at the end, and previously attached SAEs are restored to their original state. Default is True.
+            use_error_term: (Optional[bool]) If provided, will set the use_error_term attribute of all SAEs attached during this run to this value. Defaults to None.
             **model_kwargs: Keyword arguments for the model forward pass
         """
-        with self.saes(saes=saes, reset_saes_end=reset_saes_end):
+        with self.saes(
+            saes=saes, reset_saes_end=reset_saes_end, use_error_term=use_error_term
+        ):
             return self(*model_args, **model_kwargs)
 
     def run_with_cache_with_saes(
@@ -179,6 +193,7 @@ class HookedSAETransformer(HookedTransformer):
         *model_args: Any,
         saes: Union[SAE, List[SAE]] = [],
         reset_saes_end: bool = True,
+        use_error_term: Optional[bool] = None,
         return_cache_object: bool = True,
         remove_batch_dim: bool = False,
         **kwargs: Any,
@@ -200,13 +215,16 @@ class HookedSAETransformer(HookedTransformer):
             *model_args: Positional arguments for the model forward pass
             saes: (Union[HookedSAE, List[HookedSAE]]) The SAEs to be attached for this forward pass
             reset_saes_end: (bool) If True, all SAEs added during this run are removed at the end, and previously attached SAEs are restored to their original state. Default is True.
+            use_error_term: (Optional[bool]) If provided, will set the use_error_term attribute of all SAEs attached during this run to this value. Determines whether the SAE returns input or reconstruction. Defaults to None.
             return_cache_object: (bool) if True, this will return an ActivationCache object, with a bunch of
                 useful HookedTransformer specific methods, otherwise it will return a dictionary of
                 activations as in HookedRootModule.
             remove_batch_dim: (bool) Whether to remove the batch dimension (only works for batch_size==1). Defaults to False.
             **kwargs: Keyword arguments for the model forward pass
         """
-        with self.saes(saes=saes, reset_saes_end=reset_saes_end):
+        with self.saes(
+            saes=saes, reset_saes_end=reset_saes_end, use_error_term=use_error_term
+        ):
             return self.run_with_cache(  # type: ignore
                 *model_args,
                 return_cache_object=return_cache_object,  # type: ignore
@@ -255,6 +273,7 @@ class HookedSAETransformer(HookedTransformer):
         self,
         saes: Union[SAE, List[SAE]] = [],
         reset_saes_end: bool = True,
+        use_error_term: Optional[bool] = None,
     ):
         """
         A context manager for adding temporary SAEs to the model.
@@ -277,6 +296,7 @@ class HookedSAETransformer(HookedTransformer):
         Args:
             saes (Union[HookedSAE, List[HookedSAE]]): SAEs to be attached.
             reset_saes_end (bool): If True, removes all SAEs added by this context manager when the context manager exits, returning previously attached SAEs to their original state.
+            use_error_term (Optional[bool]): If provided, will set the use_error_term attribute of all SAEs attached during this run to this value. Defaults to None.
         """
         act_names_to_reset = []
         prev_saes = []
@@ -285,8 +305,9 @@ class HookedSAETransformer(HookedTransformer):
         try:
             for sae in saes:
                 act_names_to_reset.append(sae.cfg.hook_name)
-                prev_saes.append(self.acts_to_saes.get(sae.cfg.hook_name, None))
-                self.add_sae(sae)
+                prev_sae = self.acts_to_saes.get(sae.cfg.hook_name, None)
+                prev_saes.append(prev_sae)
+                self.add_sae(sae, use_error_term=use_error_term)
             yield self
         finally:
             if reset_saes_end:
