@@ -2,8 +2,9 @@
 https://github.com/ArthurConmy/sae/blob/main/sae/model.py
 """
 
+import json
 import os
-from dataclasses import dataclass
+from dataclasses import dataclass, fields
 from typing import Any, Optional
 
 import einops
@@ -11,11 +12,16 @@ import torch
 from jaxtyping import Float
 from torch import nn
 
-from sae_lens.config import LanguageModelSAERunnerConfig
+from sae_lens.config import DTYPE_MAP, LanguageModelSAERunnerConfig
 from sae_lens.sae import SAE, SAEConfig
 from sae_lens.toolkit.pretrained_sae_loaders import (
-    load_pretrained_sae_lens_sae_components,
+    handle_config_defaulting,
+    read_sae_from_disk,
 )
+
+SPARSITY_PATH = "sparsity.safetensors"
+SAE_WEIGHTS_PATH = "sae_weights.safetensors"
+SAE_CFG_PATH = "cfg.json"
 
 
 @dataclass
@@ -27,9 +33,10 @@ class TrainStepOutput:
     mse_loss: float
     l1_loss: float
     ghost_grad_loss: float
+    auxiliary_reconstruction_loss: float = 0.0
 
 
-@dataclass
+@dataclass(kw_only=True)
 class TrainingSAEConfig(SAEConfig):
 
     # Sparsity Loss Calculations
@@ -50,7 +57,8 @@ class TrainingSAEConfig(SAEConfig):
     ) -> "TrainingSAEConfig":
 
         return cls(
-            # base confg
+            # base config
+            architecture=cfg.architecture,
             d_in=cfg.d_in,
             d_sae=cfg.d_sae,  # type: ignore
             dtype=cfg.dtype,
@@ -60,6 +68,7 @@ class TrainingSAEConfig(SAEConfig):
             hook_layer=cfg.hook_layer,
             hook_head_index=cfg.hook_head_index,
             activation_fn_str=cfg.activation_fn,
+            activation_fn_kwargs=cfg.activation_fn_kwargs,
             apply_b_dec_to_input=cfg.apply_b_dec_to_input,
             finetuning_scaling_factor=cfg.finetuning_method is not None,
             sae_lens_training_version=cfg.sae_lens_training_version,
@@ -78,11 +87,19 @@ class TrainingSAEConfig(SAEConfig):
             init_encoder_as_decoder_transpose=cfg.init_encoder_as_decoder_transpose,
             scale_sparsity_penalty_by_decoder_norm=cfg.scale_sparsity_penalty_by_decoder_norm,
             normalize_activations=cfg.normalize_activations,
+            dataset_trust_remote_code=cfg.dataset_trust_remote_code,
+            model_from_pretrained_kwargs=cfg.model_from_pretrained_kwargs,
         )
 
     @classmethod
     def from_dict(cls, config_dict: dict[str, Any]) -> "TrainingSAEConfig":
-        return TrainingSAEConfig(**config_dict)
+        # remove any keys that are not in the dataclass
+        # since we sometimes enhance the config with the whole LM runner config
+        valid_field_names = {field.name for field in fields(cls)}
+        valid_config_dict = {
+            key: val for key, val in config_dict.items() if key in valid_field_names
+        }
+        return TrainingSAEConfig(**valid_config_dict)
 
     def to_dict(self) -> dict[str, Any]:
         return {
@@ -104,9 +121,11 @@ class TrainingSAEConfig(SAEConfig):
     # parameters. Maybe there's a cleaner way to do this
     def get_base_sae_cfg_dict(self) -> dict[str, Any]:
         return {
+            "architecture": self.architecture,
             "d_in": self.d_in,
             "d_sae": self.d_sae,
             "activation_fn_str": self.activation_fn_str,
+            "activation_fn_kwargs": self.activation_fn_kwargs,
             "apply_b_dec_to_input": self.apply_b_dec_to_input,
             "dtype": self.dtype,
             "model_name": self.model_name,
@@ -119,13 +138,18 @@ class TrainingSAEConfig(SAEConfig):
             "finetuning_scaling_factor": self.finetuning_scaling_factor,
             "normalize_activations": self.normalize_activations,
             "dataset_path": self.dataset_path,
+            "dataset_trust_remote_code": self.dataset_trust_remote_code,
             "sae_lens_training_version": self.sae_lens_training_version,
         }
 
 
 class TrainingSAE(SAE):
+    """
+    A SAE used for training. This class provides a `training_forward_pass` method which calculates
+    losses used for training.
+    """
 
-    cfg: TrainingSAEConfig  # type: ignore
+    cfg: TrainingSAEConfig
     use_error_term: bool
     dtype: torch.dtype
     device: torch.device
@@ -135,6 +159,15 @@ class TrainingSAE(SAE):
         base_sae_cfg = SAEConfig.from_dict(cfg.get_base_sae_cfg_dict())
         super().__init__(base_sae_cfg)
         self.cfg = cfg  # type: ignore
+
+        self.encode_with_hidden_pre_fn = (
+            self.encode_with_hidden_pre
+            if cfg.architecture != "gated"
+            else self.encode_with_hidden_pre_gated
+        )
+
+        self.check_cfg_compatibility()
+
         self.use_error_term = use_error_term
 
         self.initialize_weights_complex()
@@ -149,24 +182,33 @@ class TrainingSAE(SAE):
     def from_dict(cls, config_dict: dict[str, Any]) -> "TrainingSAE":
         return cls(TrainingSAEConfig.from_dict(config_dict))
 
-    def encode(
+    def check_cfg_compatibility(self):
+        if self.cfg.architecture == "gated":
+            assert (
+                self.cfg.use_ghost_grads is False
+            ), "Gated SAEs do not support ghost grads"
+            assert self.use_error_term is False, "Gated SAEs do not support error terms"
+
+    def encode_standard(
         self, x: Float[torch.Tensor, "... d_in"]
     ) -> Float[torch.Tensor, "... d_sae"]:
-        feature_acts, _ = self.encode_with_hidden_pre(x)
+        """
+        Calcuate SAE features from inputs
+        """
+        feature_acts, _ = self.encode_with_hidden_pre_fn(x)
         return feature_acts
 
     def encode_with_hidden_pre(
         self, x: Float[torch.Tensor, "... d_in"]
     ) -> tuple[Float[torch.Tensor, "... d_sae"], Float[torch.Tensor, "... d_sae"]]:
 
-        # move x to correct dtype
         x = x.to(self.dtype)
-
-        # handle hook z reshaping if needed.
         x = self.reshape_fn_in(x)  # type: ignore
+        x = self.hook_sae_input(x)
+        x = self.run_time_activation_norm_fn_in(x)
 
         # apply b_dec_to_input if using that method.
-        sae_in = self.hook_sae_input(x - (self.b_dec * self.cfg.apply_b_dec_to_input))
+        sae_in = x - (self.b_dec * self.cfg.apply_b_dec_to_input)
 
         # "... d_in, d_in d_sae -> ... d_sae",
         hidden_pre = self.hook_sae_acts_pre(sae_in @ self.W_enc + self.b_enc)
@@ -177,12 +219,43 @@ class TrainingSAE(SAE):
 
         return feature_acts, hidden_pre_noised
 
+    def encode_with_hidden_pre_gated(
+        self, x: Float[torch.Tensor, "... d_in"]
+    ) -> tuple[Float[torch.Tensor, "... d_sae"], Float[torch.Tensor, "... d_sae"]]:
+
+        x = x.to(self.dtype)
+        x = self.reshape_fn_in(x)  # type: ignore
+        x = self.hook_sae_input(x)
+        x = self.run_time_activation_norm_fn_in(x)
+
+        # apply b_dec_to_input if using that method.
+        sae_in = x - (self.b_dec * self.cfg.apply_b_dec_to_input)
+
+        # Gating path with Heaviside step function
+        gating_pre_activation = sae_in @ self.W_enc + self.b_gate
+        active_features = (gating_pre_activation > 0).to(self.dtype)
+
+        # Magnitude path with weight sharing
+        magnitude_pre_activation = sae_in @ (self.W_enc * self.r_mag.exp()) + self.b_mag
+        # magnitude_pre_activation_noised = magnitude_pre_activation + (
+        #     torch.randn_like(magnitude_pre_activation) * self.cfg.noise_scale * self.training
+        # )
+        feature_magnitudes = self.activation_fn(
+            magnitude_pre_activation
+        )  # magnitude_pre_activation_noised)
+
+        # Return both the gated feature activations and the magnitude pre-activations
+        return (
+            active_features * feature_magnitudes,
+            magnitude_pre_activation,
+        )  # magnitude_pre_activation_noised
+
     def forward(
         self,
         x: Float[torch.Tensor, "... d_in"],
     ) -> Float[torch.Tensor, "... d_in"]:
 
-        feature_acts, _ = self.encode_with_hidden_pre(x)
+        feature_acts, _ = self.encode_with_hidden_pre_fn(x)
         sae_out = self.decode(feature_acts)
 
         return sae_out
@@ -192,11 +265,11 @@ class TrainingSAE(SAE):
         sae_in: torch.Tensor,
         current_l1_coefficient: float,
         dead_neuron_mask: Optional[torch.Tensor] = None,
-    ) -> TrainStepOutput:  # type: ignore
+    ) -> TrainStepOutput:
 
         # do a forward pass to get SAE out, but we also need the
         # hidden pre.
-        feature_acts, _ = self.encode_with_hidden_pre(sae_in)
+        feature_acts, _ = self.encode_with_hidden_pre_fn(sae_in)
         sae_out = self.decode(feature_acts)
 
         # MSE LOSS
@@ -207,7 +280,7 @@ class TrainingSAE(SAE):
         if self.cfg.use_ghost_grads and self.training and dead_neuron_mask is not None:
 
             # first half of second forward pass
-            _, hidden_pre = self.encode_with_hidden_pre(sae_in)
+            _, hidden_pre = self.encode_with_hidden_pre_fn(sae_in)
             ghost_grad_loss = self.calculate_ghost_grad_loss(
                 x=sae_in,
                 sae_out=sae_out,
@@ -218,17 +291,40 @@ class TrainingSAE(SAE):
         else:
             ghost_grad_loss = 0.0
 
-        # SPARSITY LOSS
-        # either the W_dec norms are 1 and this won't do anything or they are not 1
-        # and we're using their norm in the loss function.
-        weighted_feature_acts = feature_acts * self.W_dec.norm(dim=1)
-        sparsity = weighted_feature_acts.norm(
-            p=self.cfg.lp_norm, dim=-1
-        )  # sum over the feature dimension
+        if self.cfg.architecture == "gated":
+            # Gated SAE Loss Calculation
 
-        l1_loss = (current_l1_coefficient * sparsity).mean()
+            # Shared variables
+            sae_in_centered = (
+                self.reshape_fn_in(sae_in) - self.b_dec * self.cfg.apply_b_dec_to_input
+            )
+            pi_gate = sae_in_centered @ self.W_enc + self.b_gate
+            pi_gate_act = torch.relu(pi_gate)
 
-        loss = mse_loss + l1_loss + ghost_grad_loss
+            # SFN sparsity loss - summed over the feature dimension and averaged over the batch
+            l1_loss = (
+                current_l1_coefficient
+                * torch.sum(pi_gate_act * self.W_dec.norm(dim=1), dim=-1).mean()
+            )
+
+            # Auxiliary reconstruction loss - summed over the feature dimension and averaged over the batch
+            via_gate_reconstruction = pi_gate_act @ self.W_dec + self.b_dec
+            aux_reconstruction_loss = torch.sum(
+                (via_gate_reconstruction - sae_in) ** 2, dim=-1
+            ).mean()
+
+            loss = mse_loss + l1_loss + aux_reconstruction_loss
+        else:
+            # default SAE sparsity loss
+            weighted_feature_acts = feature_acts * self.W_dec.norm(dim=1)
+            sparsity = weighted_feature_acts.norm(
+                p=self.cfg.lp_norm, dim=-1
+            )  # sum over the feature dimension
+
+            l1_loss = (current_l1_coefficient * sparsity).mean()
+            loss = mse_loss + l1_loss + ghost_grad_loss
+
+            aux_reconstruction_loss = torch.tensor(0.0)
 
         return TrainStepOutput(
             sae_in=sae_in,
@@ -242,6 +338,7 @@ class TrainingSAE(SAE):
                 if isinstance(ghost_grad_loss, torch.Tensor)
                 else ghost_grad_loss
             ),
+            auxiliary_reconstruction_loss=aux_reconstruction_loss.item(),
         )
 
     def calculate_ghost_grad_loss(
@@ -305,20 +402,29 @@ class TrainingSAE(SAE):
             return standard_mse_loss_fn
 
     @classmethod
-    def load_from_pretrained(  # type: ignore
+    def load_from_pretrained(
         cls,
         path: str,
         device: str = "cpu",
-        dtype: str = "float32",
+        dtype: str | None = None,
     ) -> "TrainingSAE":
 
-        config_path = os.path.join(path, "cfg.json")
-        weight_path = os.path.join(path, "sae_weights.safetensors")
+        # get the config
+        config_path = os.path.join(path, SAE_CFG_PATH)
+        with open(config_path, "r") as f:
+            cfg_dict = json.load(f)
+        cfg_dict = handle_config_defaulting(cfg_dict)
+        cfg_dict["device"] = device
+        if dtype is not None:
+            cfg_dict["dtype"] = dtype
 
-        cfg_dict, state_dict, _ = load_pretrained_sae_lens_sae_components(
-            config_path, weight_path, device, dtype
+        weight_path = os.path.join(path, SAE_WEIGHTS_PATH)
+        cfg_dict, state_dict = read_sae_from_disk(
+            cfg_dict=cfg_dict,
+            weight_path=weight_path,
+            device=device,
+            dtype=DTYPE_MAP[cfg_dict["dtype"]],
         )
-
         sae_cfg = TrainingSAEConfig.from_dict(cfg_dict)
 
         sae = cls(sae_cfg)
@@ -340,10 +446,7 @@ class TrainingSAE(SAE):
             )
             self.initialize_decoder_norm_constant_norm()
 
-        elif self.cfg.normalize_sae_decoder:
-            self.set_decoder_norm_to_unit_norm()
-
-        # Then we intialize the encoder weights (either as the transpose of decoder or not)
+        # Then we initialize the encoder weights (either as the transpose of decoder or not)
         if self.cfg.init_encoder_as_decoder_transpose:
             self.W_enc.data = self.W_dec.data.T.clone().contiguous()
         else:
