@@ -71,7 +71,7 @@ cfg = CacheActivationsRunnerConfig(
     # buffer details
     n_batches_in_buffer=4,
     store_batch_size_prompts=128,
-    normalize_activations=False,
+    normalize_activations="none",
     #
     shuffle_every_n_buffers=8,
     n_shuffles_with_last_section=1,
