@@ -119,6 +119,8 @@ def test_sae_fold_norm_scaling_factor(cfg: LanguageModelSAERunnerConfig):
     sae2 = deepcopy(sae)
     sae2.fold_activation_norm_scaling_factor(norm_scaling_factor)
 
+    assert sae2.cfg.normalize_activations == "none"
+
     assert torch.allclose(sae2.W_enc.data, sae.W_enc.data * norm_scaling_factor)
 
     # we expect activations of features to differ by W_dec norm weights.
@@ -140,7 +142,7 @@ def test_sae_fold_norm_scaling_factor(cfg: LanguageModelSAERunnerConfig):
     torch.testing.assert_close(feature_activations_2, feature_activations_1)
 
     sae_out_1 = sae.decode(feature_activations_1)
-    sae_out_2 = sae2.decode(feature_activations_2)
+    sae_out_2 = sae2.decode(feature_activations_2 * norm_scaling_factor)
 
     # but actual outputs should be the same
     torch.testing.assert_close(sae_out_1, sae_out_2)
@@ -148,6 +150,60 @@ def test_sae_fold_norm_scaling_factor(cfg: LanguageModelSAERunnerConfig):
 
 def test_sae_save_and_load_from_pretrained(tmp_path: Path) -> None:
     cfg = build_sae_cfg(device="cpu")
+    model_path = str(tmp_path)
+    sae = SAE.from_dict(cfg.get_base_sae_cfg_dict())
+    sae_state_dict = sae.state_dict()
+    sae.save_model(model_path)
+
+    assert os.path.exists(model_path)
+
+    sae_loaded = SAE.load_from_pretrained(model_path, device="cpu")
+
+    sae_loaded_state_dict = sae_loaded.state_dict()
+
+    # check state_dict matches the original
+    for key in sae.state_dict().keys():
+        assert torch.allclose(
+            sae_state_dict[key],
+            sae_loaded_state_dict[key],
+        )
+
+    sae_in = torch.randn(10, cfg.d_in, device=cfg.device)
+    sae_out_1 = sae(sae_in)
+    sae_out_2 = sae_loaded(sae_in)
+    assert torch.allclose(sae_out_1, sae_out_2)
+
+
+def test_sae_save_and_load_from_pretrained_gated(tmp_path: Path) -> None:
+    cfg = build_sae_cfg(architecture="gated", device="cpu")
+    model_path = str(tmp_path)
+    sae = SAE.from_dict(cfg.get_base_sae_cfg_dict())
+    sae_state_dict = sae.state_dict()
+    sae.save_model(model_path)
+
+    assert os.path.exists(model_path)
+
+    sae_loaded = SAE.load_from_pretrained(model_path, device="cpu")
+
+    sae_loaded_state_dict = sae_loaded.state_dict()
+
+    # check state_dict matches the original
+    for key in sae.state_dict().keys():
+        assert torch.allclose(
+            sae_state_dict[key],
+            sae_loaded_state_dict[key],
+        )
+
+    sae_in = torch.randn(10, cfg.d_in, device=cfg.device)
+    sae_out_1 = sae(sae_in)
+    sae_out_2 = sae_loaded(sae_in)
+    assert torch.allclose(sae_out_1, sae_out_2)
+
+
+def test_sae_save_and_load_from_pretrained_topk(tmp_path: Path) -> None:
+    cfg = build_sae_cfg(
+        activation_fn_str="topk", activation_fn_kwargs={"k": 30}, device="cpu"
+    )
     model_path = str(tmp_path)
     sae = SAE.from_dict(cfg.get_base_sae_cfg_dict())
     sae_state_dict = sae.state_dict()
@@ -212,3 +268,22 @@ def test_sae_get_name_returns_correct_name_from_cfg_vals() -> None:
     cfg = build_sae_cfg(model_name="test_model", hook_name="test_hook_name", d_sae=128)
     sae = SAE.from_dict(cfg.get_base_sae_cfg_dict())
     assert sae.get_name() == "sae_test_model_test_hook_name_128"
+
+
+def test_sae_move_between_devices() -> None:
+    cfg = build_sae_cfg(device="cpu")
+    sae = SAE.from_dict(cfg.get_base_sae_cfg_dict())
+
+    sae.to("meta")
+    assert sae.device == torch.device("meta")
+    assert sae.cfg.device == "meta"
+    assert sae.W_enc.device == torch.device("meta")
+
+
+def test_sae_change_dtype() -> None:
+    cfg = build_sae_cfg(device="cpu", dtype="float64")
+    sae = SAE.from_dict(cfg.get_base_sae_cfg_dict())
+
+    sae.to(dtype=torch.float16)
+    assert sae.dtype == torch.float16
+    assert sae.cfg.dtype == "torch.float16"
